32-langgraph-chat/pyproject.toml
[project]
name = "32-langgraph-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.4.0",
    "ipython>=9.9.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
